Cast.

+++ rfun.c
@@ -3286,6 +3286,6 @@ int calc_val(char *s)
 		asmerr(E_MISPAR);
 		return(0);
 	}
-	strlcpy(tmp, s, p - s + 1);
+	strlcpy(tmp, s, (size_t)(p - s + 1));
 	return (eval(tmp));
 }
